requirements.txt
click==6.7
Flask==0.12.2
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
psycopg2-binary==2.7.4
SQLAlchemy==1.2.6
elasticsearch==6.2.0
